split up assembler generator in several partial classes
